# openapi/v1.yaml
openapi: 3.1.0
info:
  title: Arbiter API
  version: 1.2.1
servers:
  - url: http://localhost:8080
paths:
...
  /v1/runs/{run_id}:
    get:
      summary: Retrieve run envelope
      parameters:
        - name: run_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Run envelope
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunEnvelope"
        "404":
          description: Run not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
...
    Run:
      type: object
      additionalProperties: false
      required:
        [run_id, request_id, requester, source, objective, environment, status, created_at, updated_at, risk_summary]
      properties:
        run_id: { type: string }
        request_id: { type: string }
        requester: { type: string }
        source: { type: string }
        objective: { type: string }
        environment: { type: string }
        status: { $ref: "#/components/schemas/RunStatus" }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        risk_summary: { type: object, additionalProperties: true }
    RunEnvelope:
      type: object
      additionalProperties: false
      required: [run, steps, approvals, permits]
      properties:
        run: { $ref: "#/components/schemas/Run" }
        steps:
          type: array
          items: { $ref: "#/components/schemas/Step" }
        approvals:
          type: array
          items: { $ref: "#/components/schemas/Approval" }
        permits:
          type: array
          items: { $ref: "#/components/schemas/ExecutionPermit" }
    StepIntent:
      type: object
      additionalProperties: false
      required: [intent_type, capability, target, risk_level, provider]
      properties:
        step_id: { type: [string, "null"] }
        client_step_id: { type: [string, "null"] }
        intent_type: { type: string }
        capability: { type: string }
        target: { type: string }
        risk_level: { type: string }
        provider: { type: string }
        metadata: { type: object, additionalProperties: true, default: {} }
    Decision:
      type: object
      additionalProperties: false
      required: [decision_id, effect, rationale, applied_policies, permit_constraints, required_approvers]
      properties:
        decision_id: { type: string }
        effect: { $ref: "#/components/schemas/DecisionEffect" }
        rationale: { type: string }
        applied_policies:
          type: array
          items: { type: string }
        permit_constraints: { type: object, additionalProperties: true }
        required_approvers:
          type: array
          items: { type: string }
    Step:
      type: object
      additionalProperties: false
      required: [step_id, run_id, status, intent, decision, created_at]
      properties:
        step_id: { type: string }
        run_id: { type: string }
        status: { $ref: "#/components/schemas/StepStatus" }
        intent: { $ref: "#/components/schemas/StepIntent" }
        decision: { $ref: "#/components/schemas/Decision" }
        permit:
          anyOf:
            - { $ref: "#/components/schemas/ExecutionPermit" }
            - { type: "null" }
        approval_id: { type: [string, "null"] }
        created_at: { type: string, format: date-time }
        updated_at: { type: [string, "null"], format: date-time }
    Approval:
      type: object
      additionalProperties: false
      required: [approval_id, run_id, step_id, status, required_approvers, reason, created_at]
      properties:
        approval_id: { type: string }
        run_id: { type: string }
        step_id: { type: string }
        status: { $ref: "#/components/schemas/ApprovalStatus" }
        required_approvers:
          type: array
          items: { type: string }
        reason: { type: string }
        created_at: { type: string, format: date-time }
        decided_at: { type: [string, "null"], format: date-time }
        decided_by: { type: [string, "null"] }
...
    ExecutionPermit:
      type: object
      additionalProperties: false
      required: [permit_id, run_id, step_id, issuer, constraints, issued_at, expires_at]
      properties:
        permit_id: { type: string }
        run_id: { type: string }
        step_id: { type: string }
        issuer: { type: string }
        constraints: { type: object, additionalProperties: true }
        issued_at: { type: string, format: date-time }
        expires_at: { type: string, format: date-time }
...
    RunStatus:
      type: string
      enum:
        - accepted
        - planning
        - waiting_for_approval
        - ready
        - running
        - blocked
        - succeeded
        - failed
        - cancelled
    StepStatus:
      type: string
      enum:
        - declared
        - evaluating
        - approval_required
        - permitted
        - executing
        - completed
        - rejected
        - failed
        - cancelled
    DecisionEffect:
      type: string
      enum: [allow, deny, require_approval]
    ApprovalStatus:
      type: string
      enum: [requested, granted, denied, cancelled]
    Error:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details:
              description: Optional structured diagnostics